+++ CHANGES.md
@@ -23,6 +23,7 @@
     - Don't override table faces by link faces [GH-716][]
     - Fix invalid italic fontification after bold markups[GH-731][]
     - Fix `markdown-live-preview-mode` fails when `eww-auto-rename-buffer` is non-nil[GH-737][]
+    - Fix to mistake to handle the line as delimiter row[GH-747][]
 
   [gh-572]: https://github.com/jrblevin/markdown-mode/issues/572
   [gh-705]: https://github.com/jrblevin/markdown-mode/issues/705
@@ -30,6 +31,7 @@
   [gh-731]: https://github.com/jrblevin/markdown-mode/issues/731
   [gh-737]: https://github.com/jrblevin/markdown-mode/issues/737
   [gh-739]: https://github.com/jrblevin/markdown-mode/issues/739
+  [gh-747]: https://github.com/jrblevin/markdown-mode/issues/747
 
 
 # Markdown Mode 2.5